Crate rattler_solve

rattler_solve is a crate that provides functionality to solve Conda environments. It currently exposes the functionality through the SolverImpl::solve function.

Modules

libsolv_c

Provides an solver implementation based on the [rattler_libsolv_c] crate.

resolvo

Provides an solver implementation based on the [resolvo] crate.

Structs

MinimumAgeConfig

Configuration for filtering packages based on their minimum age.

RepoDataIter

A helper struct that implements IntoRepoData for anything that can iterate over RepoDataRecords.

SolverTask

Represents a dependency resolution task, to be solved by one of the backends

Enums

ChannelPriority

Represents the channel priority option to use during solves.

SolveError

Represents an error when solving the dependencies for a given environment

SolveStrategy

Represents the strategy to use when solving dependencies

Traits

IntoRepoData

Defines the ability to convert a type into SolverRepoData.

SolverImpl

Represents a solver implementation, capable of solving SolverTasks

SolverRepoData

A representation of a collection of [RepoDataRecord] usable by a SolverImpl implementation.