rattler_solve

Module libsolv_c

Provides an solver implementation based on the rattler_libsolv_c crate.

Structs

LibcByteSlice

Represents an owned byte slice that was allocated using [libc::malloc] and is deallocated upon drop using [libc::free].

RepoData

Represents the information required to load available packages into libsolv for a single channel and platform combination

Solver

A Solver implemented using the libsolv library

Functions

cache_repodata

Caches the repodata as an in-memory .solv file