rattler_solve

Module resolvo

Provides an solver implementation based on the [resolvo] crate.

Structs

CondaDependencyProvider

An implement of [resolvo::DependencyProvider] that implements the ecosystem behavior for conda. This allows resolvo to solve for conda packages.

DependencyOverride

A dependency override rule.

RepoData

Represents the information required to load available packages into libsolv for a single channel and platform combination

Solver

A Solver implemented using the resolvo library

Enums

CancelReason

The reason why the solver was cancelled

NameType

Represents the type of name that is being used in the pool.

SolverMatchSpec

Wrapper around MatchSpec so that we can use it in the resolvo pool

SolverPackageRecord

Wrapper around RepoDataRecord so that we can use it in resolvo pool. Also represents a virtual package or an extra of a package.

Functions

add_extra

Adds a particular “extra” to the set of solvables

extra_version_set

Returns a version set for a particular package name and extra.