rattler_solve

Trait SolverRepoData

pub trait SolverRepoData<'a>: FromIterator<&'a RepoDataRecord> { }

A representation of a collection of [RepoDataRecord] usable by a SolverImpl implementation.

Some solvers might be able to cache the collection between different runs of the solver which could potentially eliminate some overhead. This trait enables creating a representation of the repodata that is most suitable for a specific backend.

Some solvers may add additional functionality to their specific implementation that enables caching the repodata to disk in an efficient way (see crate::libsolv_c::RepoData for an example).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.